requirements.txt
telethon==1.28.5
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
asyncio
datetime
logging
